# super categories bigint

+++ packages/graphql/src/GraphQL.queries.ts
@@ -553,7 +553,7 @@ query getPayeeById ($id: Int) {
 `;
 
 export const getSuperCategoryByIdStringQuery = `
-query getSuperCategoryById ($id: Int) {
+query getSuperCategoryById ($id: bigint) {
   item_super_categories (where: {id: {_eq: $id }}) {
     item_categories {
         id
@@ -601,7 +601,7 @@ query getItems {
 `;
 
 export const getItemsBySuperCategoryStringQuery = `
-query getItemsBySuperCategory ($id: Int, $agency_id: Int) {
+query getItemsBySuperCategory ($id: bigint, $agency_id: Int) {
   item_super_categories (where: {id: {_eq: $id }}, order_by: {id: desc}) {
       item_categories {
         items (where: {active: {_eq: true }}) {
